Fix datatype inconsistency
git-svn-id: https://ccpforge.cse.rl.ac.uk/svn/spral/branches/cpu_ssids@1102 33810a50-c353-450f-be6c-41f7241c649d

+++ src/ssids/cpu/cpu_iface.hxx
@@ -20,7 +20,7 @@ struct cpu_factor_options {
    double small;
    double u;
    int print_level;
-   int cpu_small_subtree_threshold;
+   long cpu_small_subtree_threshold;
    int cpu_task_block_size;
    PivotMethod pivot_method;
 };
